# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccessInfo": {
        "description": "La decision de acceso, para que el cliente sepa que pantalla mostrar.\n\nEs informativa: el servidor vuelve a decidir en cada llamada a un endpoint\nde juego (§2.1 de la Fase 4).",
        "properties": {
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "reason": {
            "$ref": "#/components/schemas/AccessReason"
          },
          "until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "required": [
          "granted",
          "reason",
          "until"
        ],
        "title": "AccessInfo",
        "type": "object"
      },
      "AccessReason": {
        "description": "Por que una cuenta tiene o no acceso a la mesa (`core/access.has_access`).",
        "enum": [
          "admin",
          "full",
          "invited",
          "subscription",
          "suspended",
          "consent_required",
          "expired",
          "no_access"
        ],
        "title": "AccessReason",
        "type": "string"
      },
      "AccessType": {
        "enum": [
          "none",
          "invited",
          "full"
        ],
        "title": "AccessType",
        "type": "string"
      },
...
      "AdminUserDetailResponse": {
        "description": "Una cuenta vista por el administrador (§4.2).\n\nLa suscripcion y el historial de pagos se suman en el paso 5.",
        "properties": {
          "audit": {
            "items": {
              "$ref": "#/components/schemas/AuditLogEntry"
            },
            "title": "Audit",
            "type": "array"
          },
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentResponse"
            },
            "title": "Consents",
            "type": "array"
          },
          "sessions_count": {
            "title": "Sessions Count",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user",
          "consents",
          "sessions_count",
          "audit"
        ],
        "title": "AdminUserDetailResponse",
        "type": "object"
      },
      "AdminUserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AdminUserListResponse",
        "type": "object"
      },
...
      "AuditLogEntry": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "admin_email": {
            "title": "Admin Email",
            "type": "string"
          },
          "admin_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin User Id"
          },
          "after": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "before": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "admin_user_id",
          "admin_email",
          "action",
          "target_type",
          "target_id",
          "before",
          "after",
          "reason",
          "ip",
          "created_at"
        ],
        "title": "AuditLogEntry",
        "type": "object"
      },
      "AuditLogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AuditLogListResponse",
        "type": "object"
      },
...
      "ConsentResponse": {
        "description": "Un documento que el usuario acepto, para la pagina de cuenta.",
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/LegalKind"
          },
          "legal_document_id": {
            "format": "uuid",
            "title": "Legal Document Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "legal_document_id",
          "kind",
          "title",
          "version",
          "accepted_at",
          "current"
        ],
        "title": "ConsentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LegalKind": {
        "enum": [
          "terms",
          "privacy",
          "refunds",
          "cookies"
        ],
        "title": "LegalKind",
        "type": "string"
      },
...
      "UpdateUserAccessRequest": {
        "description": "PATCH /admin/users/:id/access: otorga o retira el acceso manual.",
        "properties": {
          "access_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Expires At"
          },
          "access_type": {
            "$ref": "#/components/schemas/AccessType"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "access_type",
          "reason"
        ],
        "title": "UpdateUserAccessRequest",
        "type": "object"
      },
      "UpdateUserStatusRequest": {
        "description": "PATCH /admin/users/:id/status: suspende o reactiva la cuenta.",
        "properties": {
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "is_active",
          "reason"
        ],
        "title": "UpdateUserStatusRequest",
        "type": "object"
      },
      "UserResponse": {
        "description": "La cuenta mas su decision de acceso. Se arma con\n`api/presenters.user_response`, no con `model_validate`: `access` no es una\ncolumna, lo calcula `has_access`.",
        "properties": {
          "access": {
            "$ref": "#/components/schemas/AccessInfo"
          },
          "access_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Expires At"
          },
          "access_type": {
            "$ref": "#/components/schemas/AccessType"
          },
          "adult_confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adult Confirmed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "onboarding_completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Completed At"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "email_verified",
          "access_type",
          "access_expires_at",
          "is_active",
          "role",
          "created_at",
          "adult_confirmed_at",
          "onboarding_completed_at",
          "access"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "user",
          "admin"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Sebasanalisis API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/audit-log": {
      "get": {
        "description": "La bitacora, lo mas reciente primero. Solo lectura: no hay como borrarla.",
        "operationId": "list_audit_log_admin_audit_log_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Audit Log",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/users": {
      "get": {
        "description": "Cuentas por pagina, las mas recientes primero. Busca por correo o nombre.",
        "operationId": "list_users_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "access_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AccessType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Type"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "email_verified",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Email Verified"
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}": {
      "get": {
        "operationId": "get_user_admin_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/access": {
      "patch": {
        "description": "Otorga acceso manual (`invited` con vencimiento opcional, o `full`) o lo\nretira (`none`).",
        "operationId": "update_user_access_admin_users__user_id__access_patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserAccessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User Access",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/status": {
      "patch": {
        "description": "Suspende o reactiva una cuenta. Suspendida, entra a su cuenta pero no a la mesa.",
        "operationId": "update_user_status_admin_users__user_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User Status",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}